+++ packages/subgraph/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @smartinvoicexyz/subgraph
 
+## 0.1.23
+
+### Patch Changes
+
+- fixed subgraph
+
+## 0.1.22
+
+### Patch Changes
+
+- added support for latest subgraph + fixed useEscrowZap
+
 ## 0.1.21
 
 ### Patch Changes